fix: action equal logic
