+++ apps/builder/src/page/App/components/Actions/ActionPanel/CouchDBPanel/index.tsx
@@ -34,7 +34,6 @@ import {
 } from "@/redux/currentApp/action/couchDBAction"
 import { ResourcesData } from "@/redux/resource/resourceState"
 import { fetchResourceMeta } from "@/services/resource"
-import { updateActionContent } from "@/utils/action/calculateNewAction"
 
 type CouchDBPanelType = ActionItem<CouchDBAction<CouchDBOptionsType>>
 
@@ -46,6 +45,21 @@ const MethodSubPanelMap = {
   deleteRecord: DeleteRecordSubPanel,
   find: FindRecordSubPanel,
   getView: GetViewSubPanel,
+}
+const updateActionContent = <T extends Record<string, unknown>>(
+  obj: T,
+  path: string[],
+  value: string | boolean,
+) => {
+  if (path.length === 1) {
+    const key = path[0]
+    return { ...obj, [key]: value } as T
+  }
+  const [currentKey, ...restKeys] = path
+  const currentObj = obj[currentKey] as Record<string, unknown>
+  const updatedObj = updateActionContent(currentObj, restKeys, value) as T
+
+  return { ...obj, [currentKey]: updatedObj } as T
 }
 
 export const CouchDBPanel: FC = () => {
